Adding KeyPair Config for more options
